notifications_api/worker/requirements.txt
aiosmtplib==2.0.0
alembic==1.9.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
black==22.12.0
blinker==1.5
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==3.2
django-environ==0.8.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-mail==1.2.4
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.0.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.3
pika==1.3.1
platformdirs==2.6.2
pydantic==1.9.0
PyJWT==2.0.1
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sniffio==1.3.0
urllib3==1.26.13
uvicorn==0.20.0
vine==5.0.0
Werkzeug==2.2.2